pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-gateway"
version = "0.1.0"
description = "Internal API gateway for vendor API access"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # HTTP client
    "httpx>=0.28.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Redis
    "redis[hiredis]>=5.2.0",
    # Configuration
    "pydantic-settings>=2.6.0",
    # JWT / Auth
    "python-jose[cryptography]>=3.3.0",
    # Logging
    "structlog>=24.4.0",
    # Observability
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-exporter-prometheus>=0.50b0",
    "prometheus-client>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "testcontainers[redis,postgres]>=4.9.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gateway"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "testcontainers[redis,postgres]>=4.9.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]
